requirements.txt
appdirs==1.4.4
arrow==1.1.0
black==21.5b0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.17.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
